Add files via upload
Added a FSM diagram for part3 in the part3 folder. Created a getJunit method in the utilities class along with a call to it under WMethod class. Created a new package named Proj3 with 2 classes JamesBond and JBondTest. JamesBond has a method named bondRegex which returns a boolean based a pattern made using regular expressions. JamesBondTest is the JUnit tests created from the WMethod's getJUnit method.